+++ internal/acceptance_tests/blocks/logging_grafanacloudlogs_compute.tf
@@ -1,0 +1,11 @@
+resource "fastly_service_logging_grafanacloudlogs" "test" {
+  service_id = fastly_service_compute.test.id
+  version    = {{.SERVICE_VERSION}}
+  name       = "{{.LOGGING_GRAFANACLOUDLOGS_NAME}}"
+  url        = "https://test123.grafana.net"
+  user       = "123456"
+  index      = "{\"label\": \"value\"}"
+  authentication = {
+    token = "test-grafanacloudlogs-token"
+  }
+}