+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 
 ### ENHANCEMENTS:
 
+- feat(grafana_cloud_logs): add support for Grafana Cloud Logs ([#92](https://github.com/fastly/terraform-provider-fastly-beta/pull/92))
 - feat(ngwaf_workspace_templated_signal_rules): moved Templated Signal Rules to the dedicated `fastly_ngwaf_workspace_templated_signal_rules` data source ([#90](https://github.com/fastly/terraform-provider-fastly-beta/pull/90))
 - feat(user_service_authorization): add `fastly_user_service_authorization` resource for granting user permissions on a service ([#94](https://github.com/fastly/terraform-provider-fastly-beta/pull/94))
 
 ### BUG FIXES:
+
+- fix(cloudfiles_logging): ensure that the `format` attribute can't be set to null ([#93](https://github.com/fastly/terraform-provider-fastly-beta/pull/93))
 
 ### Dependencies: